add better support for branching files from other projects
When a file is branched from another project outside of the one vss2git is migrating, we cannot retrieve all the information about that file. Without these changes, the code tries to get the latest version of the original file but always retrieves the first version as it doesn't have any information about the past versions of the file from the other project. This also builds a descriptive commit message for share actions instead of leaving it blank.